REFACTOR: primitive.py
